Time: 1 ms (95.03%) | Memory: 47.1 MB (20.17%) - LeetSync

+++ 3846-minimum-operations-to-make-array-sum-divisible-by-k/minimum-operations-to-make-array-sum-divisible-by-k.java
@@ -1,6 +1,6 @@
 class Solution {
     public int minOperations(int[] nums, int k) {
-        int s=0;
+        long s=0;
         for(int i=0;i<nums.length;i++)
         {
